requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.7
fastjsonschema==2.15.0
Flask==1.1.2
Flask-Compress==1.9.0
fsspec==2021.4.0
future==0.18.2
gbqschema-converter==1.2.1
gcsfs==2021.4.0
google-api-core==1.26.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
libcst==0.3.18
MarkupSafe==1.1.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
plotly==4.14.3
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.8
py==1.10.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3
